Relocate tworker queue override logic to run_bot
