+++ src/clusterfuzz/_internal/base/tasks/__init__.py
@@ -366,29 +366,7 @@ def get_preprocess_task():
   return task
 
 
-@memoize.wrap(memoize.FifoInMemory(1))
-def _get_tworker_queue_override() -> str:
-  """Gets the tworker queue override from environment or metadata."""
-  queue_override = environment.get_value('OVERRIDE_QUEUE')
-  if queue_override:
-    return queue_override
-
-  if not compute_metadata.is_gce():
-    return ""
-
-  try:
-    queue_override = compute_metadata.get('instance/attributes/override_queue')
-    if queue_override:
-      return queue_override.strip()
-  except exceptions.RequestException as e:
-    if not (isinstance(e, exceptions.HTTPError) and
-            e.response.status_code == 404):
-      logs.warning(f'Error fetching override_queue metadata: {e}')
-
-  return ""
-
-
-def tworker_get_task():
+def tworker_get_task(override_queue: str = None):
   """Gets a task for a tworker to do."""
   assert environment.is_tworker()
   # TODO(metzman): Pulling tasks is relatively expensive compared to
@@ -396,10 +374,9 @@ def tworker_get_task():
   # queue that is probably empty) to do a single preprocess. Investigate
   # combining preprocess and postprocess queues and allowing pulling of
   # multiple messages.
-  queue_override = _get_tworker_queue_override()
-  if queue_override:
-    logs.info(f'Overriding default queue to {queue_override}')
-    return get_regular_task(queue=queue_override)
+  if override_queue:
+    logs.info(f'Overriding default queue to {override_queue}')
+    return get_regular_task(queue=override_queue)
 
   if random.random() < .5:
     # Pick either one with equal probability so we don't hurt the
