+++ src/clusterfuzz/_internal/tests/core/bot/startup/run_bot_test.py
@@ -204,3 +204,90 @@ class ScheduleUtaskMainsTest(unittest.TestCase):
 
     self.mock.RemoteTaskGate.return_value.create_utask_main_jobs.assert_called_once(
     )
+
+
+class TworkerGetTaskTest(unittest.TestCase):
+  """Tests for tworker_get_task."""
+
+  def setUp(self):
+    helpers.patch(self, [
+        'clusterfuzz._internal.system.environment.get_value',
+        'clusterfuzz._internal.system.environment.is_tworker',
+        'clusterfuzz._internal.base.tasks.get_regular_task',
+        'clusterfuzz._internal.base.tasks.get_postprocess_task',
+        'clusterfuzz._internal.base.tasks.get_preprocess_task',
+        'clusterfuzz._internal.base.tasks.random.random',
+        'clusterfuzz._internal.google_cloud_utils.compute_metadata.is_gce',
+        'clusterfuzz._internal.google_cloud_utils.compute_metadata.get',
+    ])
+    self.mock.is_tworker.return_value = True
+    self.mock.is_gce.return_value = False
+
+  def test_override_queue_set(self):
+    """Test that tworker_get_task returns a task from the override queue."""
+    self.mock.get_value.return_value = 'override_queue'
+    override = run_bot._get_tworker_queue_override(__memoize_force__=True)  # pylint: disable=protected-access
+    mock_task = mock.Mock()
+    self.mock.get_regular_task.return_value = mock_task
+
+    result = taskslib.tworker_get_task(override_queue=override)
+
+    self.assertEqual(result, mock_task)
+    self.mock.get_regular_task.assert_called_once_with(queue='override_queue')
+
+  def test_override_queue_not_set(self):
+    """Test that tworker_get_task falls back to random choice when override queue is not set."""
+    self.mock.get_value.return_value = None
+    override = run_bot._get_tworker_queue_override(__memoize_force__=True)  # pylint: disable=protected-access
+    self.mock.random.return_value = 0.4
+    mock_task = mock.Mock()
+    self.mock.get_postprocess_task.return_value = mock_task
+
+    result = taskslib.tworker_get_task(override_queue=override)
+
+    self.assertEqual(result, mock_task)
+    self.mock.get_postprocess_task.assert_called_once()
+
+  def test_override_queue_empty(self):
+    """Test that tworker_get_task falls back to random choice when override queue is empty."""
+    self.mock.get_value.return_value = ''
+    override = run_bot._get_tworker_queue_override(__memoize_force__=True)  # pylint: disable=protected-access
+    self.mock.random.return_value = 0.6
+    mock_task = mock.Mock()
+    self.mock.get_preprocess_task.return_value = mock_task
+
+    result = taskslib.tworker_get_task(override_queue=override)
+
+    self.assertEqual(result, mock_task)
+    self.mock.get_preprocess_task.assert_called_once()
+
+  def test_override_queue_from_metadata_success(self):
+    """Test that tworker_get_task returns a task from the override queue in metadata."""
+    self.mock.get_value.return_value = None
+    self.mock.is_gce.return_value = True
+    self.mock.get.return_value = ' metadata_override '
+    override = run_bot._get_tworker_queue_override(__memoize_force__=True)  # pylint: disable=protected-access
+    mock_task = mock.Mock()
+    self.mock.get_regular_task.return_value = mock_task
+
+    result = taskslib.tworker_get_task(override_queue=override)
+
+    self.assertEqual(result, mock_task)
+    self.mock.get_regular_task.assert_called_once_with(
+        queue='metadata_override')
+
+  def test_override_queue_from_metadata_exception(self):
+    """Test that tworker_get_task falls back to original behavior when metadata fetch fails."""
+    from requests import exceptions
+    self.mock.get_value.return_value = None
+    self.mock.is_gce.return_value = True
+    self.mock.get.side_effect = exceptions.RequestException("Failed")
+    override = run_bot._get_tworker_queue_override(__memoize_force__=True)  # pylint: disable=protected-access
+    self.mock.random.return_value = 0.4
+    mock_task = mock.Mock()
+    self.mock.get_postprocess_task.return_value = mock_task
+
+    result = taskslib.tworker_get_task(override_queue=override)
+
+    self.assertEqual(result, mock_task)
+    self.mock.get_postprocess_task.assert_called_once()
